+++ src/importlinter/application/use_cases.py
@@ -368,6 +368,10 @@ def _get_built_in_contract_types() -> list[tuple[str, type[Contract]]]:
                 "independence: importlinter.contracts.independence.IndependenceContract",
                 "protected: importlinter.contracts.protected.ProtectedContract",
                 "acyclic_siblings: importlinter.contracts.acyclic_siblings.AcyclicSiblingsContract",
+                (
+                    "unused_public_symbols: importlinter.contracts.unused_public_symbols"
+                    ".UnusedPublicSymbolsContract"
+                ),
             ],
         )
     )